requirements.txt
flask==3.1.0
flask-login==0.6.3
flask-wtf==1.2.2
flask-session==0.8.0
supabase==2.12.0
bcrypt==4.2.1
cryptography==44.0.0
gunicorn==23.0.0
python-dotenv==1.0.1
requests==2.32.3
redis==5.0.0
anthropic==0.49.0
PyJWT==2.9.0
Pillow==11.1.0
openai==1.77.0
sentry-sdk[flask]==2.20.0
markdown==3.7
